Add Sphinx HTML github actions workflow
On pushes to the `main` branch, create `_build` and `_static` folders
under `docs`. These are used by sphinx to generate HTML files
for use on GitHub Pages:

https://jcassady.github.io/llava-benchmark/

It is expected that `.rst` updatedfiles from the `sphinx-apidoc`
command are present in the `docs` folder so the workflow
can build the html pages and deploy them to the `gh-pages`
branch.

Resolves PR #12